[Feat] 디테일 페이지 ui 수정 중
