+++ src/App.js
@@ -6,7 +6,7 @@ import Main from "./pages/Main";
 import View from "./pages/View";
 import Share from "./pages/Share";
 import Together from "./pages/Together";
-
+import PerformanceDetail from "./pages/PerformanceDetail";
 function App() {
   return (
     <>
@@ -17,6 +17,10 @@ function App() {
         <Route path="/view" element={<View />} />
         <Route path="/share" element={<Share />} />
         <Route path="/together" element={<Together />} />
+        <Route
+          path="/share/performance_detail"
+          element={<PerformanceDetail />}
+        />
       </Routes>
       <Footer />
     </>
